Added option to indent and not indent XML
When marshalling the signed XML there can be no indentation used for
readability since it invalidates the signature.  Added indent boolean
parameter to the XMLUtil.marshalToStream method to enable indentation
when sending to standard out and to not use indentation when
marshalling to the stream for parsing and validating the XML signature.